+++ recalc-vscode/src/frontend/external.d.ts
@@ -1,0 +1,9 @@
+/** Data passed to index.js  */
+
+// as specified in package.json
+declare const config: {
+  locale: "EN_US" | "ZH_CN";
+  serverUri: string;
+};
+
+declare const data: SheetDocument;