+++ recalc-vscode/CHANGELOG.md
@@ -1,7 +1,12 @@
 # Revision history for recalc-vscode
 
-## 0.0.3 -- 2025-02-07
+## 0.0.4 -- 2025-02-07
 
 * Initial version: create project structure for a custom editor.
 * Add generic rpc implementation, client implementation and logging facilities.
 * Integrate type-safe json-rpc communication between client and server.
+* Add simple UI code using [Univer core/sheet-api][univer-sheet-api].
+
+<!-- References -->
+
+  [univer-sheet-api]: https://docs.univer.ai/en-US/guides/sheets/features/core/sheet-api